+++ cofi-share/src/main/AndroidManifest.xml
@@ -3,4 +3,15 @@
 
     <uses-permission android:name="android.permission.POST_NOTIFICATIONS" />
     <uses-permission android:name="android.permission.VIBRATE" />
+
+    <application>
+        <receiver
+                android:name=".timer.notification.TimerActionBroadcastReceiver"
+                android:exported="false">
+            <intent-filter>
+                <action android:name="COFI_TIMER_ACTION_BROADCAST" />
+            </intent-filter>
+        </receiver>
+    </application>
+
 </manifest>
